requirements.txt
cffi==1.17.1
cryptography==44.0.3
greenlet==3.2.1
iniconfig==2.1.0
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
pytest==8.3.5
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
